+++ sdk/core/Azure.Core/src/Pipeline/HttpClientTransport.Request.cs
@@ -3,8 +3,11 @@
 
 using System;
 using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
+using System.Linq;
 using System.Net.Http;
 using System.Runtime.InteropServices;
+using System.ServiceModel.Rest.Core;
 using System.ServiceModel.Rest.Core.Pipeline;
 
 namespace Azure.Core.Pipeline
@@ -14,8 +17,19 @@ namespace Azure.Core.Pipeline
     /// </summary>
     public partial class HttpClientTransport : HttpPipelineTransport, IDisposable
     {
-        // TODO: is there a way to still do this with this private?  Come back to this.
-        internal sealed class HttpClientTransportRequest : HttpPipelineRequest
+        internal static bool TryGetPipelineRequest(Request request, out PipelineRequest? pipelineRequest)
+        {
+            if (request is RequestAdapter requestAdapter)
+            {
+                pipelineRequest = requestAdapter.PipelineRequest;
+                return true;
+            }
+
+            pipelineRequest = null;
+            return false;
+        }
+
+        private sealed class HttpClientTransportRequest : HttpPipelineRequest
         {
             private RequestUriBuilder? _uriBuilder;
 
@@ -78,6 +92,88 @@ namespace Azure.Core.Pipeline
 #else
                 httpRequest.Properties[name] = value;
 #endif
+            }
+        }
+
+        private class RequestAdapter : Request
+        {
+            private readonly HttpClientTransportRequest _request;
+
+            public RequestAdapter(HttpClientTransportRequest request)
+            {
+                _request = request;
+            }
+
+            internal PipelineRequest PipelineRequest => _request;
+
+            public override RequestMethod Method
+            {
+                get => RequestMethod.Parse(_request.Method);
+                set => _request.Method = value.Method;
+            }
+
+            public override RequestUriBuilder Uri
+            {
+                get => _request.UriBuilder;
+                set => _request.UriBuilder = value;
+            }
+
+            public override RequestContent? Content
+            {
+                get => (RequestContent?)_request.Content;
+                set => _request.Content = value;
+            }
+
+            public override void Dispose() => _request.Dispose();
+
+            protected internal override void AddHeader(string name, string value)
+                => _request.Headers.Add(name, value);
+
+            protected internal override bool ContainsHeader(string name)
+                => _request.Headers.TryGetHeader(name, out _);
+
+            protected internal override IEnumerable<HttpHeader> EnumerateHeaders()
+            {
+                _request.Headers.TryGetHeaders(out IEnumerable<MessageHeader<string, object>> headers);
+                foreach (var header in headers)
+                {
+                    yield return new HttpHeader(header.Name, GetHeaderValueString(header.Name, header.Value));
+                }
+            }
+
+            // TODO: avoid copying this?
+            private static string GetHeaderValueString(string name, object value) => value switch
+            {
+                string s => s,
+                List<string> l => string.Join(",", l),
+                _ => throw new InvalidOperationException($"Unexpected type for header {name}: {value?.GetType()}")
+            };
+
+            protected internal override bool RemoveHeader(string name)
+                => _request.Headers.Remove(name);
+
+            protected internal override bool TryGetHeader(string name, [NotNullWhen(true)] out string? value)
+                => _request.Headers.TryGetHeader(name, out value);
+
+            protected internal override bool TryGetHeaderValues(string name, [NotNullWhen(true)] out IEnumerable<string>? values)
+            {
+                // TODO: we can optimize this, come back to it
+                _request.Headers.TryGetHeaders(out IEnumerable<MessageHeader<string, object>> headers);
+
+                var header = headers.Where(h => h.Name == name);
+                if (header == null || !header.Any())
+                {
+                    values = null;
+                    return false;
+                }
+
+                values = header.First().Value switch
+                {
+                    string s => new string[] { s },
+                    List<string> l => l,
+                    _ => throw new InvalidOperationException($"Unexpected type for header {name}: {header.First().Value?.GetType()}")
+                };
+                return true;
             }
         }
     }
