+++ sdk/core/Azure.Core/src/Pipeline/HttpClientTransport.cs
@@ -64,13 +64,7 @@ namespace Azure.Core.Pipeline
         /// <inheritdoc />
         [EditorBrowsable(EditorBrowsableState.Never)]
         public sealed override Request CreateRequest()
-        {
-            HttpClientTransportRequest requestImpl = new HttpClientTransportRequest();
-            RequestAdapter request = new RequestAdapter(requestImpl);
-            //requestImpl.UriBuilder = request.Uri;
-
-            return request;
-        }
+            => new RequestAdapter(new HttpClientTransportRequest());
 
         /// <inheritdoc />
         public override void Process(HttpMessage message)
