+++ sdk/core/Azure.Core/src/Pipeline/HttpWebRequestTransport.cs
@@ -47,16 +47,16 @@ namespace Azure.Core.Pipeline
         /// <inheritdoc />
         public override void Process(HttpMessage message)
         {
-            ProcessInternal(message, false).EnsureCompleted();
+            ProcessSyncOrAsync(message, false).EnsureCompleted();
         }
 
         /// <inheritdoc />
         public override async ValueTask ProcessAsync(HttpMessage message)
         {
-            await ProcessInternal(message, true).ConfigureAwait(false);
+            await ProcessSyncOrAsync(message, true).ConfigureAwait(false);
         }
 
-        private async ValueTask ProcessInternal(HttpMessage message, bool async)
+        private async ValueTask ProcessSyncOrAsync(HttpMessage message, bool async)
         {
             var request = CreateRequest(message.Request);
